Clean-up and migration of Text.Xml.Conduit.Elements
`ppElements' is only used by Network.Xmpp.Xep.InbandRegistration, so I
moved it there for the time-being.

As `elements' is only used in Stream.hs, I moved it there.

`parseElement' is not used anywhere, and was removed completely.

`renderElement' (and its local `elementToEvents' function) is used by
Stream.hs and Concurrent.hs, so I moved it to Utilties.hs. (We could
argue for a separate `Elements' module, but right now that seems
a bit thin. `openElementToEvents' is used by both `elementToEvents'
and `renderOpenElement', so I put both `openElementToEvents' and
`renderOpenElement' in Utilities.hs as well.

`compressNodes' and `streamName' were made where-local to `elements'.

The types were moved to Types.hs.