Add Read and Show instances for Jid
Here's a hack with `Read' and `Show' instances for Jid, as discussed.
I didn't have the time or patience to make it elegant, but at least it
does the job.

I'm calling it `jid' instead of `toJid' for the time-being since
`toJid' is already used in the stream configuration record.

Fixes #24.