add run tests
We lacked a test suite that even just tried connecting to a server and
bouncing a few stanzas back and forth. The "runtests" suite does just
that. It connects to 2 accounts on a server (set in a config
file ~/.pontarius-xmpp-tests/pontarius-xmpp-tests.conf) and tries
passing some messages and IQ stanzas back and forth. This should clear
the way for more interesting scenarios.